revise endCondition logic so timer only delets when timer exists

+++ utils/endConditions.js
@@ -37,7 +37,11 @@ const cleanUp = async (usersArray) => {
       if(!user.isActive){
         // if(isBotRoleHigher({ member: user.member }))
         restoreNickname(user, user.member);
-        user.botTimerMessage.delete();
+
+        if(user.userSetTimer){
+          user.botTimerMessage.delete();
+
+        }
 	
         if(user.mode !== 'shame') {
           console.log('exited early');
@@ -51,7 +55,10 @@ const cleanUp = async (usersArray) => {
 			
         // if(isBotRoleHigher({ member: user.member })) 
         restoreNickname(user, user.member);
-        user.botTimerMessage.delete();
+        if(user.userSetTimer){
+          user.botTimerMessage.delete();
+
+        }
 	
         if(user.mode !== 'shame') {
           console.log('timer up');
